Fixed Mantis 5265 were command region-remove also removed objects from database

+++ OpenSim/Region/Application/OpenSimBase.cs
@@ -557,7 +557,10 @@ namespace OpenSim
                 SceneManager.TrySetCurrentScene("..");
             }
 
-            scene.DeleteAllSceneObjects();
+            if (cleanup)
+            {
+                scene.DeleteAllSceneObjects();
+			}
             SceneManager.CloseScene(scene);
             ShutdownClientServer(scene.RegionInfo);
